#65 refactoring first part
